Reimplemented Analysis class to clear up regression code now that it's conciderably more streamlined.
